adk-chatbot/pyproject.toml
[project]
name = "adk-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.2.10",
    "amadeus>=12.0.0",
    "click>=8.2.1",
    "dotenv>=0.9.9",
    "google-adk[eval]>=1.5.0",
    "google-cloud-aiplatform>=1.101.0",
    "pinecone>=7.3.0",
]
